requirements.txt
PyQt6>=6.4.0
pyserial>=3.5
pyinstaller>=6.0.0

